library/requirements.txt
django
djangorestframework
markdown
django-filte